UPSTREAM: drivers/ipmi: Add Supermicro OEM commands
Add a new driver for OEM commands and select it from x11-lga1151-series.

The driver communicates the BIOS version and date to the BMC using OEM
commands. The command should be supported on all X11 series mainboards,
but might work with older BMC, too.

Tested on X11SSH-TF:
The BIOS version strings are updated on boot and are visible in the
BMC web UI.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Dossym Nurmukhanov <dossym@google.com>
Original-Commit-Id: 09cdeba60643cb85f91b233211617e95167a64e5
Original-Change-Id: I51c22f83383affb70abb0efbcdc33ea925b5ff9f
Original-Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Original-Signed-off-by: Christian Walter <christian.walter@9elements.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38002
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Michael Niewöhner <foss@mniewoehner.de>
Change-Id: I1bf676dc95868585eb752798554f6bd34c9e116e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2604728
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>