UPSTREAM: mb/hp/8770w: Transform into variant
Since this board does not have integrated graphics, do not install the
INT15 handler if it is not selected in Kconfig.

NOTE: Since cmos options are not very flexible, this board ends up with
a spurious gfx_uma_size option. Other than that, everything is the same.

Tested with BUILD_TIMELESS=1, binary does not change when ignoring the
cmos options.

BUG=none
BRANCH=none
TEST=none

Change-Id: If985adde137245fd15444e1713dcf40b984604c2
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 533e4399f10940796fedd6056ada60426bcc673f
Original-Change-Id: I2ebcfd5160773bf98a3d23e797a89e290063d112
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38094
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2009117
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>