UPSTREAM: mb/hp/snb_ivb_laptops: Switch to overridetree setup
NOTE: The ME interface was disabled on folio_9470m and revolve_810_g1.
It is assumed that they were ported while the ME was in an abnormal
state (usually due to me_cleaner usage), and that it should be enabled.
In any case, the MEI device is hidden if the ME fails to boot already.

BUG=none
BRANCH=none
TEST=none

Change-Id: I16ed07b1df3b6928ceff3b2a8d7f3860121db426
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 42d300533e6388d0c31d700ec5ea29d21e0216cd
Original-Change-Id: Ibf32a034653946f49f72a2c19c41a4033964ef83
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38097
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2009120
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>