UPSTREAM: vc/amd/agesa/[...]/Config: Avoid out-of-bounds warnings
The memNTrainFlowControl array is generating Coverity warnings
in multiple places in code where it attempts to write to index 1.
The array is defined as either 2 elements or 1 of NULL depending
on #if (AGESA_ENTRY_INIT_POST == TRUE). This is likely a false
alarm from Coverity (memory should not be training outside of a
POST), but adding a second NULL element for the
AGESA_ENTRY_INIT_POST == FALSE case. Tested on Lenovo G505s.

BUG=none
BRANCH=none
TEST=none

Change-Id: I6bc2d2226035acb0265b85f63e3cb5bb63c8d142
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a608dd80d514a8036954f6a0e5d19645e6ba3d9d
Original-Change-Id: Iaebe0830471e1854d6191c69cdaa552f900ba7a6
Original-Signed-off-by: Joe Moore <awokd@danwin1210.me>
Original-Found-by: Coverity CID 1357451, 1357452, 1357453
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38176
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2091293
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>