UPSTREAM: drivers/spi/spi_flash: separate out protection ops
Put the write protection into its own object. This allows
for easier future reuse of objects in future consolidation
patches. It's also possible to eliminate the code implmementing
these in the future if the platform doesn't require it. For now
leave current behavior as-is.

The names of the callbacks were shortened as they are now in
the spi_flash_protection_ops object which is a new field in the
spi_flash object.

BUG=none
BRANCH=none
TEST=none

Change-Id: I8e090cfeac5f3542ffc7fdbb392c4d0be802edec
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f584f19efc58ba3c80a462024ae680320b17b9ce
Original-Change-Id: I2fec4e4430709fcf3e08a55dd36583211c035c08
Original-Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38376
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2009029
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>