UPSTREAM: mb/lenovo/x201: Support undock button on X200 ultrabase
Only the ACPI code needs to be extended, as smihandler.c already
supported it. The _Q50 Method is just _Q18 with changed name.

On Linux, pressing the undock button does nothing, so the only safe
way to undock is to press Fn+F9. With this patch, when the undock
button is pressed, the green LED lights up, and undocking is safe.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ida84200eb2c5488fda7b3124bc004624d30dd0f9
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d76947f13b760d876fdfb675bac992cb4275e1c1
Original-Change-Id: Iaaecad031bb1f39dd1a778d0c8eaea6bce9e0f57
Original-Signed-off-by: Maciej Matuszczyk <maccraft123mc@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38446
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Mimoja <coreboot@mimoja.de>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2010886
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>