UPSTREAM: cpu/x86: Make MP init timeout configurable
The current MP init timeout is hardcoded as 1s. To support
platform with many cpus, the timeout needs to be adjusted.
The number of cpus is calculated as:
number of sockets * number of cores per socket *
  number of threads per core

How long the timeout should be set to, is heuristic.
It needs to be set long enough to ensure reboot stability,
but not unreasonable so that real failures can be detected
soon enough, especially for smaller systems.

This patch sets timeout to be minimum as 1 second, while each
cpu adds 0.1 second.

BUG=none
BRANCH=none
TEST=none

Change-Id: I8cbf844339362cfdb2372fa9c0967614f1c2f824
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6ec322ec7c27005d64a907076188b09a7cf09aa7
Original-Signed-off-by: Jonathan Zhang <jonzhang@fb.com>
Original-Signed-off-by: Reddy Chagam <anjaneya.chagam@intel.com>
Original-Change-Id: Ibc079fc6aa8641d4ac8d8e726899b6c8d055052e
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38546
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: David Hendricks <david.hendricks@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2037476
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>