UPSTREAM: soc/intel: Add Intel Xeon Scalable Processor support
This patch adds support for Intel Xeon-SP.

This patch is developed and verified with Skylake Scalable
Processor, which is a processor in Xeon-SP family. The code
is expected to be reusable for future geneations of Xeon-SP
processors, and will be updated with smaller targeted
patches accordingly, to add support for additional Xeon-SP
processors, to add features, and to improve the code base.

The Skylake-SP FSP is based on FSP 2.0. The FSP is a
proof-of-concept build. The binary is not shared in public,
when this patch is upstreamed.

BUG=none
BRANCH=none
TEST=none

Change-Id: I5ae6572d77b72a655edb46240fdc09a96af2aa39
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8f89549d3c7d41643337662947cfdb2329bd030b
Original-Signed-off-by: Jonathan Zhang <jonzhang@fb.com>
Original-Signed-off-by: Reddy Chagam <anjaneya.chagam@intel.com>
Original-Tested-by: johnny_lin@wiwynn.com
Original-Change-Id: Idc9c3bee17caf8b4841f0bc190cb1aa9d38fc23e
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38548
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: David Hendricks <david.hendricks@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2091358
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>