UPSTREAM: soc/amd/picasso: Add APOB NV back for non-S3
New information indicates the PSP expects the APOB NV region
populated for all types of boot, and this is not a feature only
used for S3.  Switch over to using the MRC_CACHE flash region.

Remove the Kconfig symbols for the APOB_NV base and size.  Override
the MRC_CACHE_SETTINGS_CACHE_SIZE to ensure the default maintains the
minimum required size.  Use the generated (or mainboard-specified)
fmap.fmd file as an input for amdfwtool and properly match the
flash region.

Change the original naming for the APOB destination, which matched the
PSP spec's field name, to PSP_APOB_DESTINATION.  This should be more
intuitive for a source code reader.  The APOB address is the location
in DRAM where the PSP puts its output block.

BUG=b:147042464, b:153675914
TEST=Boot trembyle

Change-Id: I18cf95dbf383ea1d2a8b228712195343a14d6a8c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b768723c72b326b24703a8e67764c995711050d9
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38703
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Original-Change-Id: Ia5ba8646deec2bd282df930f471738723063eef8
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2080375
Original-Original-Change-Id: I972d66f1817f86ff0b689f011c0c44c3fe7c8ef7
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2053312
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Change-Id: I4550766ece462b65a6bfe6f1b747343e08e53fe5
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2215985
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>