UPSTREAM: src/device: Add option to look at revision in option roms
AMD's Family 17h SOCs have the same vendor and device IDs for
their graphics blocks, but need different video BIOSes.  The
only difference is the revision number.

Add a Kconfig option that allows us to add the revision number
of the graphics device to the PCI option rom saved in CBFS.

Because searching CBFS takes a non-trivial amount of time,
only enable the option if it's needed.  If it's not used, or
if nothing matches, the check will fall through and search for
an option rom with no version.

BUG=b:145817712
TEST=With surrounding patches, loads dali vbios

Change-Id: I3e361d537d1a46b4c44d1784d67b6e23ffa33045
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6cf5d2f0d3869d0c53e95b81ea01fa5b8a60cbd0
Original-Change-Id: Icb610a2abe7fcd0f4dc3716382b9853551240a7a
Original-Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/2013181
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Tested-by: Martin Roth <martinroth@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39792
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2121252
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>