UPSTREAM: util/amdfwtool: Clarify APOB NV requirements
Relocate the first size check.  This was automatically continuing
and not looking for the caller incorrectly passing a destination.

New information indicates that the APOB_NV should always be present
in the system.  Augment the missing size check to inferring whether
a missing size is valid, as in the case of older products, or truly
missing when it's needed.

BUG=none
BRANCH=none
TEST=none

Change-Id: Idc3b60ca960819cad5335f2906a41432f60f240b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c4a8c48b2f70d56c7c318f4ce24a467a1d708ef5
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Change-Id: I51f5333de4392dec1478bd84563c053a508b9e9e
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38690
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2083540
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>