UPSTREAM: google/octopus: Disable bootblock console
The GLK bootblock seems(?) to be hard limited to 32KB and some Octopus
variants are so close to that that they only have 0.5KB left. This is
blocking development of new core features, so let's disable the
bootblock console to gain a couple of KB back (like we already did on
RK3288).

There are probably other opporunities for code size reduction here (e.g.
it seems that almost half(!) of that whole bootblock size is taken up by
devicetree.cb structures), but I'm not familiar enough with the platform
to dig into them.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib9c4e1c36c138546b55d61f2c56b10223e22e9ed
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 3f3b4d5d74c4c9d4f01c7035c8626032cb9a6b56
Original-Change-Id: I05b4ecf5abef7307e3d0a81db04a745ff3da0c42
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38521
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2022701
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>