UPSTREAM: drivers/spi/spi_flash: remove spi flash names
The names of each spi flash cause quite a bit of bloat in the text
size of each stage/program. Remove the name entirely from spi flash
in order to reduce overhead. In order to pack space as closely as
possible the previous 32-bit id and mask were split into 2 16-bit
ids and masks.

On Chrome OS build of Aleena there's a savings of >2.21KiB in each
of verstage, romstage, and ramstage.

BUG=none
BRANCH=none
TEST=none

Change-Id: I4f846e961a7ac8349e54daa30caf5ca2e87c2e8a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: fc7b953366b776aa3eaaf0539af06086facda14e
Original-Change-Id: Ie98f7e1c7d116c5d7b4bf78605f62fee89dee0a5
Original-Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38380
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2025590
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>