UPSTREAM: cpu: Allow to configure microcode at pre-defined address
FSP-T takes microcode pointer and location parameters, and FSP-T is
invoked before CAR is set-up and before memory is trained. So it is not
possible to modify supplied microcode pointer in runtime. Because of
that we have to hardcode the pointer in bootblock.

Also, current FSP-T on Xeons require microcode (it is not optional).
Reasons for that are currently unclear and are being investigated.

However for the present time we need to be able to add microcode at a
certain offset so FSP-T can be used.

TEST=test on OCP TiogaPass board, as well as out-of-tree CPU/board

Change-Id: I09d27b32e8153f0e377622f4f3503ede21d2ce19
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d1f7c6f2867ad7ede4af36ed42860a206538c550
Original-Change-Id: I6c02601a7ac64078e556e2032baeccaf27f77da2
Original-Signed-off-by: Andrey Petrov <anpetrov@fb.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38640
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Philipp Deppenwiese <zaolin.daisuki@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2062594
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>