UPSTREAM: mb/lenovo/t440p: Enable dGPU on Lenovo T440P
Enable the dGPU on the Lenovo T440P.  It uses the same code (roughly) of
the T430S.  By default, it is set to be disabled however it can be
enabled via the nvram option enable_dual_graphics. Removed hybrid graphics
options too as they are not valid for the T440p. Tested on a T440P with
Ubuntu 18.04.4 with Kernel 5.3.0-29 (successful). Tested on same machine
with Windows 10 1909 (machine check exception bluescreen).

BUG=none
BRANCH=none
TEST=none

Change-Id: Ie214aa824095a20ab139e487d7359cca2bc2935c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f8a13d5a2292c08515e4970d3f9bf5a9d0bee7d1
Original-Change-Id: Idf8c2c0d1ae34bda8736448d3e350396e3cf7a93
Original-Signed-off-by: Chris Morgan <macromorgan@hotmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38723
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2046281
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>