UPSTREAM: nb/intel/haswell/peg: Add PEG driver stub
This is a port of https://review.coreboot.org/c/coreboot/+/22337 to the
Haswell northbridge.  This code is necessary to support the dGPU of the
t440p. Code was cut and pasted from Sandy Bridge with vendor IDs updated
to the correct Haswell values.  Tested on t440p with dGPU on Ubuntu
18.04.4 with 5.3.0-28 kernel. Without patches dmesg reports Nouveau is
unable to read the VBIOS of the dGPU as it has an invalid checksum (I
checked that the ROM in CBFS is correct). With this patch DRM works
correctly with both the Nouveau driver and the Nvidia proprietary
driver. Windows 10 1909 also tested but generates bluescreen once GPU
driver is loaded.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ice4fa04b05fdd157ab9024a189793fa0644989b8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5e5e789f9b115f7fb1e7c453cdf20df088ac893d
Original-Change-Id: Ie5f089fb6fd774e6c61f4f9281e2945bd44edf27
Original-Signed-off-by: Chris Morgan <macromorgan@hotmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38743
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2091310
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>