UPSTREAM: util/ifdtool: add --output flag
Add an optional commandline flag to define the filename of the resulting
output file. If this flag is not defined, it will behave like before by
using the old filename with a ".new" suffix.

With this additional flag it is not necessary to move the output file at
build-time, and the stdout print "Writing new image to <filename>" makes
more sense in the build context.

BUG=none
BRANCH=none
TEST=none

Change-Id: If0460b0f4b49e77555e5d75877f321318a52ede6
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: dc1596c8c8246b22dfb77a7214745782bab02d9e
Original-Change-Id: I824e94e93749f55c3576e4ee2f7804d855fefed2
Original-Signed-off-by: Marcello Sylvester Bauer <sylv@sylv.io>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38828
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2084462
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>