UPSTREAM: Revert "mb/google/hatch: Override CPU flex ratio"
This reverts commit a017e5fb3dda5ea6bbc94ee15b2e981eeaa2d918.

Reason for revert: The extra reset in the FSP after the flex ratio is changed causes recovery reasons to be lost. There are some vboot changes that recently landed that could help with this issue, but for now, we are working on a new AU image for Kohaku and this is causing our automated testing to fail.

BUG=none
BRANCH=none
TEST=none

Change-Id: I232639b95229cadc958fa4b4342044eaef974059
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ca15430bf7d43c2ba0c0684f745f206ab5f70670
Original-Change-Id: Ic38b390842e2a533033587b3247b7c8d982b1dff
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38324
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2060621
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>