UPSTREAM: cpu/x86/smm: Convert C++ style comment
Originally, this patch made 'BIOS' uppercase in the referenced comment
and converted the C++ style to be consistent with the remainder of
the function.  Somewhere, the 'BIOS' became uppercase creating a merge
conflict.

Now this CL converts the C++ style to be consistent with the remainder
of the comments.

BUG=none
BRANCH=none
TEST=none

Change-Id: I5a68cb32ead285815165648033294a45d1dd354b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 17f0f0118853f3f35897cf5aa14661eb9ec08ac5
Original-Signed-off-by: Eugene D. Myers <edmyers@tycho.nsa.gov>
Original-Change-Id: I85d78b5e08a7643c3d87e3daf353d6b3ba8d306b
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38854
Original-Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2070279
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>