UPSTREAM: cpu/x86/name: Make name.c file available in romstage
In this patch, name.c file that includes the function definition for
fill_processor_name which is used by the report_cpu_info function is been
made available in romstage.

This is done to facilitate the report_platform_info to be called from
romstage, as the intention is to move the report_platform_info to romstage
for all SOC's due to the bootblock size constraint.

BUG=None
TEST=Build and boot APL, GLK and CNL platforms.

Change-Id: I02efa56ed106aa17b0bc95b91cd5b641b6e1dc80
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5bf7ffbe0814528436cfef5bdca00c1ef7f584d1
Original-Change-Id: Ifd6d4b80c2e07d02adaed676a56efeb6fb704552
Original-Signed-off-by: Usha P <usha.p@intel.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/38940
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Subrata Banik <subrata.banik@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2062821
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>