UPSTREAM: vc/amd/agesa/f14: Fix function return type
F14GetNbCofVidUpdate() is declared elsewhere to be of type
F_CPU_IS_NBCOF_INIT_NEEDED, which is supposed to return a boolean value
(not an AGESA status). This is fixed in the corresponding f15tn and
f16kb code, so apply the same change here. This fixes a compiler error
when using LTO.

BUG=none
BRANCH=none
TEST=none

Change-Id: I0acdf243f20cf2e861388020913264487b1cb6bf
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 365f52eb1f3692133bae8cc7e0cac5dc81d9de92
Original-Change-Id: Ifc44e2c0467f8bd1f537b5a69c501ba51053d3d9
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/39013
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2154674
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>