UPSTREAM: vc/amd/agesa/f15tn,f16kb: Fix array types
These variables are declared to be arrays of MICROCODE_PATCHES_4K (which
is a struct containing a UINT8 array). However, the actual definitions
of these arrays ignore the wrapping struct and just use the underlying
UINT8 arrays directly, which causes a compiler error when using LTO
because of the type mismatch. Fix the type declaration so that it
matches.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ie8d3bf1af0e29e5d2739d1d245ab5c40540c04eb
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9e3e49234d30dadd5888b389ed572a2faac1c6af
Original-Change-Id: I6bef27507092fe72fe2f836c427ebb2c19009e78
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40436
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2154684
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>